requirements.txt
flask>=2.3.0
psycopg[binary,pool]>=3.1.0
databricks-sdk>=0.18.0
requests>=2.31.0
pandas>=1.5.0
PyYAML>=6.0 